+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,9 @@
+
+import Foundation
+
+// вопрос
+struct QuizQuestion {
+  let image: String         // строка с названием фильма, совпадает с названием картинки в Assets
+  let text: String          // строка с вопросом о рейтинге фильма
+  let correctAnswer: Bool   // правильный ответ на вопрос true-false
+}
