+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,10 @@
+
+import Foundation
+
+
+// вью модель для состояния "Результат квиза"
+struct QuizResultsViewModel {
+  let title: String             // строка с заголовком алерта
+  let text: String              // строка с текстом о количестве набранных очков
+  let buttonText: String        // текст для кнопки алерта
+}
