+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,10 @@
+
+import UIKit
+
+
+// вью модель для состояния "Вопрос показан"
+struct QuizStepViewModel {
+  let image: UIImage            // картинка с афишей фильма
+  let question: String          // вопрос о рейтинге квиза
+  let questionNumber: String    // строка с порядковым номером этого вопроса (ex. "1/10")
+}
